requirements.txt
python-telegram-bot
openai
flask
pydantic
agents